sound: Update all channels on I/O writes
So far, only channel 3 is updated immediately when its I/O registers are
written.  This is because I assumed one had to re-initialize (NRx4.7) a
channel to make it load updated values; but that is not the case.
Hence, immediately update all channels when their respective I/O
registers are written.