Add WASM/web-sys target
This requires adding a src/lib.rs, which is effectively empty on all
targets but wasm32.  It is a bit of a shame that we therefore create an
empty library file, but I cannot get Cargo to skip building the library
by default, so that is how it is now.

To build the WASM packages, run:
$ wasm-pack build --target web

(Or just ./build-wasm.sh.)

I really hope I didn't break anything in the non-wasm build too badly...